+++ src/presenters/collection-item.js
@@ -6,7 +6,7 @@ import Markdown from 'Components/text/markdown';
 import Text from 'Components/text/text';
 import { TeamAvatar, UserAvatar } from 'Components/images/avatar';
 import Image from 'Components/images/image';
-import Loader from 'Components/loaders/loader';
+import Loader from 'Components/loader';
 
 import CollectionOptionsContainer from './pop-overs/collection-options-pop';
 import { TeamLink, UserLink, CollectionLink, ProjectLink } from './includes/link';
