+++ src/presenters/includes/team-analytics.js
@@ -6,7 +6,7 @@ import sampleAnalytics, { sampleAnalyticsTime } from 'Curated/sample-analytics';
 
 import Text from 'Components/text/text';
 import SegmentedButtons from 'Components/buttons/segmented-buttons';
-import Loader from 'Components/loaders/loader';
+import Loader from 'Components/loader';
 
 import TeamAnalyticsTimePop from '../pop-overs/team-analytics-time-pop';
 import TeamAnalyticsProjectPop from '../pop-overs/team-analytics-project-pop';
